move the AgentClientUDP definition out of the utils package to improve SOC
